+++ engine/src/systems/GizmoSystem.cpp
@@ -36,7 +36,7 @@ void GizmoSystem::init(World *world)
 
 void GizmoSystem::update(const Input &input, const Time &time)
 {
-    for (int i = 0; i < mWorld->getNumberOfComponents<Camera>(); i++)
+    for (size_t i = 0; i < mWorld->getNumberOfComponents<Camera>(); i++)
     {
         Camera *camera = mWorld->getComponentByIndex<Camera>(i);
